requirements.txt
cycler==0.10.0
kiwisolver==1.0.1
matplotlib==3.0.2
numpy==1.15.4
Pillow==5.4.0
pkg-resources==0.0.0
pyparsing==2.3.0
python-dateutil==2.7.5
six==1.12.0
